Don't let the user delete/cut all the pages (bug 2021828)
And dispatch an event only when the context menu is displayed.

+++ web/pdf_thumbnail_viewer.js
@@ -141,10 +141,6 @@ class PDFThumbnailViewer {
 
   #scrollableContainerHeight = 0;
 
-  #previousStates = {
-    hasSelectedPages: false,
-  };
-
   #statusLabel = null;
 
   #statusBar = null;
@@ -236,6 +232,29 @@ class PDFThumbnailViewer {
         }
       });
 
+      this.container.addEventListener(
+        "contextmenu",
+        e => {
+          this.eventBus.dispatch("editingstateschanged", {
+            source: this,
+            details: {
+              thumbnailId:
+                parseInt(
+                  e.target
+                    .closest(".thumbnailImageContainer")
+                    ?.parentElement.getAttribute("page-number")
+                ) ?? -1,
+              hasSelectedPages: !!this.#selectedPages?.size,
+              canDeletePages: this.#canDelete(),
+            },
+          });
+        },
+        {
+          signal: abortSignal,
+          passive: true,
+        }
+      );
+
       this.#undoButton?.addEventListener("click", this.#undo.bind(this));
       this.#undoCloseButton?.addEventListener(
         "click",
@@ -252,24 +271,6 @@ class PDFThumbnailViewer {
     );
     this.#resetView();
     this.#addEventListeners();
-  }
-
-  /**
-   * Update the different possible states of this manager, e.g. is there
-   * something to copy, paste, ...
-   * @param {Object} details
-   */
-  #dispatchUpdateStates(details) {
-    const hasChanged = Object.entries(details).some(
-      ([key, value]) => this.#previousStates[key] !== value
-    );
-
-    if (hasChanged) {
-      this.eventBus.dispatch("editingstateschanged", {
-        source: this,
-        details: Object.assign(this.#previousStates, details),
-      });
-    }
   }
 
   #scrollUpdated() {
@@ -759,6 +760,11 @@ class PDFThumbnailViewer {
     });
   }
 
+  #canDelete() {
+    const size = this.#selectedPages?.size || 0;
+    return size > 0 && size < this._thumbnails.length;
+  }
+
   #togglePasteMode(enable) {
     this.#isInPasteMode = enable;
     if (enable) {
@@ -808,6 +814,10 @@ class PDFThumbnailViewer {
   }
 
   #cutPages() {
+    if (!this.#canDelete()) {
+      return;
+    }
+
     this.#isCut = true;
     this.#copyPages(false);
     this.#deletePages(/* type = */ "cut");
@@ -839,10 +849,11 @@ class PDFThumbnailViewer {
   }
 
   #deletePages(type = "delete") {
-    const selectedPages = this.#selectedPages;
-    if (selectedPages.size === 0) {
+    if (!this.#canDelete()) {
       return;
     }
+
+    const selectedPages = this.#selectedPages;
     if (type === "delete") {
       this.#updateStatus("delete");
     }
@@ -868,14 +879,10 @@ class PDFThumbnailViewer {
   }
 
   #updateMenuEntries() {
-    this.#manageSaveAsButton.disabled =
-      this.#manageDeleteButton.disabled =
-      this.#manageCopyButton.disabled =
-      this.#manageCutButton.disabled =
-        !this.#selectedPages?.size;
-    this.#dispatchUpdateStates({
-      hasSelectedPages: !!this.#selectedPages?.size,
-    });
+    const size = this.#selectedPages?.size || 0;
+    this.#manageSaveAsButton.disabled = this.#manageCopyButton.disabled = !size;
+    this.#manageDeleteButton.disabled = this.#manageCutButton.disabled =
+      !this.#canDelete();
   }
 
   #toggleMenuEntries(enable) {
@@ -884,9 +891,6 @@ class PDFThumbnailViewer {
       this.#manageCopyButton.disabled =
       this.#manageCutButton.disabled =
         !enable;
-    this.#dispatchUpdateStates({
-      hasSelectedPages: false,
-    });
   }
 
   #updateStatus(type) {
@@ -1102,16 +1106,6 @@ class PDFThumbnailViewer {
         this.#computeThumbnailsPosition();
       }
     });
-    this.container.addEventListener("focusout", () => {
-      this.#dispatchUpdateStates({
-        hasSelectedPages: false,
-      });
-    });
-    this.container.addEventListener("focusin", () => {
-      this.#dispatchUpdateStates({
-        hasSelectedPages: !!this.#selectedPages?.size,
-      });
-    });
     this.container.addEventListener("keydown", e => {
       const { target } = e;
       const isCheckbox =
@@ -1218,6 +1212,7 @@ class PDFThumbnailViewer {
       if (
         e.button !== 0 || // Skip right click.
         this.#isInPasteMode ||
+        this._thumbnails.length === 1 ||
         !isNaN(this.#lastDraggedOverIndex) ||
         !draggedImage.classList.contains("thumbnailImageContainer")
       ) {
